music-poller/requirements.txt
requests
python-mpd2
